better structure of all branch functions
